+++ connectivity/FEATURE_BLE/include/ble/GattServer.h
@@ -20,14 +20,14 @@
 #ifndef MBED_GATT_SERVER_H__
 #define MBED_GATT_SERVER_H__
 
-#include "CallChainOfFunctionPointersWithContext.h"
+#include "ble/common/CallChainOfFunctionPointersWithContext.h"
 
 #include "ble/common/ble/GattService.h"
 #include "ble/common/ble/GattAttribute.h"
 #include "ble/common/ble/GattServerEvents.h"
 #include "ble/common/ble/GattCallbackParamTypes.h"
 
-#include "ble/common/ble/blecommon.h"
+#include "ble/common/blecommon.h"
 #include "ble/Gap.h"
 #include "SecurityManager.h"
