+++ connectivity/FEATURE_BLE/include/ble/common/BLERoles.h
