+++ connectivity/FEATURE_BLE/include/ble/common/Bounded.h
