+++ connectivity/FEATURE_BLE/include/ble/common/SafeBool.h
