+++ connectivity/FEATURE_BLE/include/ble/common/SafeEnum.h
