+++ connectivity/FEATURE_BLE/include/ble/common/ble/GattAttribute.h
@@ -19,8 +19,8 @@
 #ifndef MBED_GATT_ATTRIBUTE_H__
 #define MBED_GATT_ATTRIBUTE_H__
 
-#include "ble/common/ble/UUID.h"
-#include "ble/common/ble/BLETypes.h"
+#include "ble/common/UUID.h"
+#include "ble/common/BLETypes.h"
 
 /**
  * @addtogroup ble
