+++ connectivity/FEATURE_BLE/include/ble/common/ble/GattService.h
@@ -19,7 +19,7 @@
 #ifndef MBED_GATT_SERVICE_H__
 #define MBED_GATT_SERVICE_H__
 
-#include "ble/common/ble/UUID.h"
+#include "ble/common/UUID.h"
 #include "ble/common/ble/GattCharacteristic.h"
 
 /**
