+++ connectivity/FEATURE_BLE/include/ble/common/blecommon.h
