+++ connectivity/FEATURE_BLE/include/ble/gap/AdvertisingDataSimpleBuilder.h
@@ -21,7 +21,7 @@
 
 #include "platform/Span.h"
 
-#include "ble/common/ble/UUID.h"
+#include "ble/common/UUID.h"
 #include "ble/gap/AdvertisingDataBuilder.h"
 
 namespace ble {
