+++ connectivity/FEATURE_BLE/include/ble/gap/AdvertisingParameters.h
@@ -21,9 +21,9 @@
 
 #include <algorithm>
 
-#include "ble/common/ble/BLETypes.h"
-#include "ble/common/ble/blecommon.h"
-#include "ble/common/ble/SafeEnum.h"
+#include "ble/common/BLETypes.h"
+#include "ble/common/blecommon.h"
+#include "ble/common/SafeEnum.h"
 
 namespace ble {
