+++ connectivity/FEATURE_BLE/include/ble/gap/ConnectionParameters.h
@@ -21,7 +21,7 @@
 
 #include "mbed_assert.h"
 
-#include "ble/common/ble/BLETypes.h"
+#include "ble/common/BLETypes.h"
 
 namespace ble {
