+++ connectivity/FEATURE_BLE/source/generic/GattClientImpl.h
@@ -21,11 +21,11 @@
 
 #include "ble/BLE.h"
 
-#include "CallChainOfFunctionPointersWithContext.h"
+#include "ble/common/CallChainOfFunctionPointersWithContext.h"
 #include <algorithm>
 
 
-#include "ble/common/ble/blecommon.h"
+#include "ble/common/blecommon.h"
 #include "ble/common/ble/GattAttribute.h"
 #include "ble/common/ble/ServiceDiscovery.h"
 #include "ble/common/ble/CharacteristicDescriptorDiscovery.h"
