+++ connectivity/FEATURE_BLE/source/generic/SecurityManagerImpl.h
@@ -20,15 +20,15 @@
 #define IMPL_SECURITY_MANAGER_H_
 
 #include <stdint.h>
-#include "CallChainOfFunctionPointersWithContext.h"
+#include "ble/common/CallChainOfFunctionPointersWithContext.h"
 #include "platform/Callback.h"
 
-#include "ble/common/ble/BLETypes.h"
-#include "ble/common/ble/blecommon.h"
+#include "ble/common/BLETypes.h"
+#include "ble/common/blecommon.h"
 #include "ble/Gap.h"
 
 #include "ble/common/ble/GapTypes.h"
-#include "ble/common/ble/BLETypes.h"
+#include "ble/common/BLETypes.h"
 #include "SecurityDb.h"
 #include "source/pal/PalConnectionMonitor.h"
 #include "source/pal/PalSigningMonitor.h"
