+++ connectivity/FEATURE_BLE/source/pal/PalSigningMonitor.h
@@ -19,7 +19,7 @@
 #ifndef MBED_BLE_SIGNING_EVENT_MONITOR
 #define MBED_BLE_SIGNING_EVENT_MONITOR
 
-#include "ble/common/ble/BLETypes.h"
+#include "ble/common/BLETypes.h"
 
 namespace ble {
